feat(components): extract Custom Footer to be an external component - extract Custom Footer from the vanilla bundle and make a separate component as an external package so it could be use by Aurelia-Slickgrid/Angular-Slickgrid/... - also extract the Binding Helper & Service into a seperate package as well to that we can use it within the Custom Footer package without the need to also require vanilla-grid-bundle - this external component will still be a required dependency of the vanilla-grid-bundle, it just brings the flexibility of it being available to other framework
