+++ packages/binding/src/__tests__/index.spec.ts
@@ -1,0 +1,7 @@
+import * as services from '../index';
+
+describe('Testing Services entry point', () => {
+  it('should have multiple service entries defined', () => {
+    expect(services).toBeTruthy();
+  });
+});
