+++ packages/custom-footer-component/README.md
@@ -1,30 +1,24 @@
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![TypeScript](https://img.shields.io/badge/%3C%2F%3E-TypeScript-%230074c1.svg)](http://www.typescriptlang.org/)
 [![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
-[![npm](https://img.shields.io/npm/v/@slickgrid-universal/composite-editor-component.svg?color=forest)](https://www.npmjs.com/package/@slickgrid-universal/composite-editor-component)
-[![npm](https://img.shields.io/npm/dy/@slickgrid-universal/composite-editor-component?color=forest)](https://www.npmjs.com/package/@slickgrid-universal/composite-editor-component)
+[![npm](https://img.shields.io/npm/v/@slickgrid-universal/custom-footer-component.svg?color=forest)](https://www.npmjs.com/package/@slickgrid-universal/custom-footer-component)
+[![npm](https://img.shields.io/npm/dy/@slickgrid-universal/custom-footer-component?color=forest)](https://www.npmjs.com/package/@slickgrid-universal/custom-footer-component)
 
 [![Actions Status](https://github.com/ghiscoding/slickgrid-universal/workflows/CI%20Build/badge.svg)](https://github.com/ghiscoding/slickgrid-universal/actions)
 [![Cypress.io](https://img.shields.io/badge/tested%20with-Cypress-04C38E.svg)](https://www.cypress.io/)
 [![jest](https://jestjs.io/img/jest-badge.svg)](https://github.com/facebook/jest)
 [![codecov](https://codecov.io/gh/ghiscoding/slickgrid-universal/branch/master/graph/badge.svg)](https://codecov.io/gh/ghiscoding/slickgrid-universal)
 
-## Composite Editor Component
-#### @slickgrid-universal/composite-editor-component
+## Custom Footer Component
+#### @slickgrid-universal/custom-footer-component
 
-Vanilla Bundle implementation of a Composite Editor Modal Window which can do the following
-- Create
-- Clone (allows you to clone & edit a row, it's like a copy+edit in a single action)
-- Update
-- Mass Update Changes
-- Mass Selection Changes (similar to Mass Update but only for the selected items/rows)
+Vanilla Implementation of a Custom Footer Component that can optionally be shown at the bottom of the grid. The default behavior is to show metrics on the right side of the footer while the left side could optionally show row selection count (when that feature is enabled). The default texts for metrics & row selection (when enabled) also have a set of locales that allows you to change the text with current locale. Lastly, you could also choose to provide your own custom text/html in both left and/or right side.
 
 ### Internal Dependencies
 - [@slickgrid-universal/common](https://github.com/ghiscoding/slickgrid-universal/tree/master/packages/common)
 
 ### External Dependencies
-- [assign-deep](https://github.com/jonschlinkert/assign-deep) to deeply assign Object properties/values
 - [DOM Purify](https://github.com/cure53/DOMPurify) to sanitize HTML text
 
 ### Installation
-Follow the instruction provided in the main [README](https://github.com/ghiscoding/slickgrid-universal#installation).
+Follow the instruction provided in the main [README](https://github.com/ghiscoding/slickgrid-universal#installation)
