+++ packages/custom-footer-component/src/index.ts
@@ -1,0 +1,1 @@
+export * from './slick-footer.component';
