+++ packages/vanilla-bundle/README.md
@@ -14,6 +14,15 @@
 
 Vanilla Bundle implementation (no framework, plain TypeSript implementation). This package does what other framework would do, that is to make all the features usable in 1 bundle so that it could then be used by other Apps/Projects, for example we use this bundle in our SalesForce (with Lighning Web Component) App and it requires plain ES6 JavaScript which this bundle also produce (for that there's a `dist-grid-bundle-zip` folder which will zip the ES6 `dist` folder which we then import in SalesForce).
 
+### Internal Dependencies
+- [@slickgrid-universal/common](https://github.com/ghiscoding/slickgrid-universal/tree/master/packages/common)
+- [@slickgrid-universal/custom-footer-component](https://github.com/ghiscoding/slickgrid-universal/tree/master/packages/custom-footer-component)
+- [@slickgrid-universal/empty-warning-component](https://github.com/ghiscoding/slickgrid-universal/tree/master/packages/empty-warning-component)
+
+### External Dependencies
+- [DOM Purify](https://github.com/cure53/DOMPurify) to sanitize HTML text
+- [whatwg-fetch](https://github.com/whatwg/fetch) - Fetch Standard
+
 ### Installation
 This Vanilla Bundle is used in our SalesForce implementation (since it requires plain ES6) and is also used by the standalone `webpack-demo-vanilla-bundle` which serves for demo purposes.
