+++ packages/vanilla-bundle/package.json
@@ -48,6 +48,7 @@
   "dependencies": {
     "@slickgrid-universal/common": "^0.14.1",
     "@slickgrid-universal/composite-editor-component": "^0.14.1",
+    "@slickgrid-universal/custom-footer-component": "^0.14.1",
     "@slickgrid-universal/empty-warning-component": "^0.14.1",
     "@slickgrid-universal/excel-export": "^0.14.1",
     "@slickgrid-universal/text-export": "^0.14.1",
