+++ packages/vanilla-bundle/src/services/index.ts
@@ -1,5 +1,3 @@
-export * from './binding.service';
-export * from './binding.helper';
 export * from './eventPubSub.service';
 export * from './resizer.service';
 export * from './textExport.service';
